Create the initial screen
Create the main heading of the web page and the container for the
etch-a-sketch screen squares in the index.html file.

Create a script that creates a number of 16x16 squares of initial size
on the sketch screen.

Give the web page an initial style in the style.css file and use flexbox
to convert the squares created on the sketch screen by the script into a
grid.