Set up a “hover” effect on the sketch screen
Create a new class in the style.css file with a certain background
color.

Add an event listener in the script for each square of the grid
so that, every time the mouse is moved over one of the squares, the new
class created in CSS is assigned to them and their background color
is changed.