Add new feature: random color hover
Remove the ".hovered-square" class selector and the flex property of the
".square" class selector in the style.css file to manipulate the
"hovered" square color and the square proportion with JavaScript.

Create a new function that gets random rgb values to randomize the
hovered square colors.

Change the name of the function that creates grids.

Add the square proportion functionality when a grid is created.

Add the random color functionality when a square is hovered.

Change the name of the variable of the new squares-per-side value
entered by the user.